Move Schema and SyntaxTree to ast::
